DB설계 변경 완료 회원가입,로그인 로직 테스트 완료

+++ src/main/java/com/dasolsystem/core/entity/ApprovalRequest.java
@@ -16,7 +16,7 @@ public class ApprovalRequest {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "request_id", updatable = false, nullable = false)
-    private Integer requestId;
+    private Long requestId;
 
     /** 요청자 (기안자 회원) */
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
